fix compilation issues

+++ VL53L0X.cpp
@@ -302,10 +302,10 @@ void VL53L0X::writeReg32Bit(uint8_t reg, uint32_t value)
 {
   Wire.beginTransmission(address);
   Wire.write(reg);
-  Wire.write((value >> 24) & 0xFF); // value highest byte
-  Wire.write((value >> 16) & 0xFF);
-  Wire.write((value >>  8) & 0xFF);
-  Wire.write( value        & 0xFF); // value lowest byte
+  Wire.write((byte)(value >> 24) & 0xFF); // value highest byte
+  Wire.write((byte)(value >> 16) & 0xFF);
+  Wire.write((byte)(value >>  8) & 0xFF);
+  Wire.write((byte) value        & 0xFF); // value lowest byte
   last_status = Wire.endTransmission();
 }
